requirements.txt
requests>=2.4.3
apache-libcloud>=1.4.0
Pillow>=3.2.0